requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# AI APIs
openai>=1.12.0              # DALL-E 3 API
anthropic>=0.18.0           # Claude API

# Image processing
Pillow>=10.2.0              # Image manipulation
opencv-python>=4.9.0        # Advanced image processing

# PDF generation
reportlab>=4.0.0            # PDF export
pypdf>=3.17.0               # PDF manipulation

# Data handling
pydantic>=2.5.0             # Data validation
jsonschema>=4.20.0          # JSON validation

# Utilities
requests>=2.31.0            # HTTP requests
tqdm>=4.66.0                # Progress bars
colorama>=0.4.6             # Colored terminal output

# Development
pytest>=7.4.0               # Testing
black>=23.12.0              # Code formatting
flake8>=7.0.0               # Linting
